+++ CHANGES.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [NEW] Added functionality to retrieve a previous version of a document.
 
 - [NEW] Add new view parameters, `stable` and `update`, as keyword arguments to `get_view_result`.
 - [FIXED] Case where an exception was raised after successful retry when using `doc.update_field`.